generate day04 and move its solution and tests over
Also remove old src dir
