+++ 2024/rust/day04/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "day04"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+clap.workspace = true
+nom.workspace = true
