+++ 2024/rust/day04/inputs/example
