+++ 2024/rust/day04/inputs/puzzle.md
