--- 2024/rust/src/day3/example
@@ -1,1 +1,0 @@
-xmul(2,4)%&mul[3,7]!@^do_not_mul(5,5)+mul(32,64]then(mul(11,8)mul(8,5))
