--- 2024/rust/src/day3/example2
@@ -1,1 +1,0 @@
-xmul(2,4)&mul[3,7]!^don't()_mul(5,5)+mul(32,64](mul(11,8)undo()?mul(8,5))
